uppdated Root.py, media.py and script.js. Also added authentication.py
Added the file authentication.py. In Root.py I added the functionality
of authentication.py, that is, it checks if a user is authorized to use
the media  player. Also changed things in script.js to accomidate this.
In media.py I commented away the play(nr) function since it messed with
the normal play.